Master Yi Jungle
This script is mainly for aster Yi jungler.
It has auto-smite, auto-dodge, visual range of spells and attacks and
much more.

NOTE: I did not create this script.  I am uploading this scripts as a
backup repository if the creator would remove it or anything, it's still
here for you to grab if you liked it.
It's also for my own database for scripts I like so I don't need to find
all of them all over again if I remove the Loader.